NavBar extracted to another file
